feat: 메인페이지 디자인 변경 및 캡쳐 처리
- 디바운스 할 때 캡쳐 후 프론트 반환 예정
- 메인페이지 디자인 간단히 구성
