+++ client/src/components/MainTitle/MainTitle.jsx
@@ -7,7 +7,7 @@ const TITLE_INFO = {
   DESCRIPTION: (
     <span>
       Recordly는 "대학생들을 위한 필기 솔루션"이라는 비전으로 <br />
-      사용자의 메모에 대해 자동 태깅 및 추천을 해줌으로써
+      사용자의 메모에 대해 자동 태깅 및 메모 추천을 해줌으로써
       <br />
       사용자에게 새로운 경험을 줄 수 있도록 생태계를 만들어 나가고 있습니다.
     </span>
