+++ client/src/components/TagContainer/TagContainer.module.scss
@@ -1,6 +1,3 @@
 .TagContainer {
-  height: 80vh;
-  border: 1px solid gray;
-
-  overflow: scroll;
+  overflow: hidden;
 }
