+++ client/src/components/WorkspaceList/WorkspaceList.module.scss
@@ -1,6 +1,6 @@
 .WorkspaceList {
   &__container {
-    width: 8rem; // 144px
+    width: 10rem; // 160px
     display: flex;
     flex-direction: column;
     align-items: center;
