+++ client/src/pages/main/Main.module.scss
@@ -1,1 +1,17 @@
+.Main {
+  // background-color: black;
+  height: 100vh;
+  width: 100%;
 
+  &__container {
+    width: 90%;
+    &__tagList {
+      max-width: 20rem;
+      min-width: 15rem;
+      border: 0.5px solid rgb(186, 186, 186);
+    }
+
+    &__mainDashboard {
+    }
+  }
+}
