+++ server/package.json
@@ -12,6 +12,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "capture-website": "^3.0.1",
     "concurrently": "^7.3.0",
     "connect-redis": "^6.1.3",
     "cookie-parser": "^1.4.6",
